Better temp file naming

+++ isis_server/routes/start.py
@@ -1,9 +1,11 @@
 from flask_expects_json import expects_json
 from flask import request, jsonify, current_app
 from os import remove as remove_file
+from os.path import splitext, exists as path_exists
 from requests import get as req_get
-from tempfile import NamedTemporaryFile
+from urllib.parse import urlparse
 
+from ..config import Config
 from ..input_validation import get_json_schema
 from ..logger import get_logger
 
@@ -16,12 +18,16 @@ def download(url):
     """
     Downloads the file at url to a temporary file & returns the temporary file
     """
-    with req_get(url, stream=True) as req, NamedTemporaryFile(mode='wb', delete=False) as temp_file:
+    input_path = urlparse(url).path
+    _, ext = splitext(input_path)
+    output_file = Config.get_tmp_file(ext)
+
+    with req_get(url, stream=True) as req, open(output_file, mode='wb') as temp_file:
         req.raise_for_status()
         for req_chunk in req.iter_content(READ_CHUNK_SZ):
             temp_file.write(req_chunk)
 
-        return temp_file.name
+    return output_file
 
 
 @expects_json(get_json_schema())
@@ -30,6 +36,7 @@ def post_start():
 
     # Either err or output_file will be set, but not both
     err = None
+    temp_file = None
     output_file = None
 
     try:
@@ -40,10 +47,12 @@ def post_start():
         logger.debug("{} downloaded to {}".format(input_file, temp_file))
 
         output_file = current_app.s3_client.upload(temp_file)
-        remove_file(temp_file)
 
     except Exception as e:
         err = str(e)
+
+    if temp_file is not None and path_exists(temp_file):
+        remove_file(temp_file)
 
     return jsonify({
         "to": output_file,
